pyproject.toml
[tool.poetry]
name = "bookmanagement"
version = "0.1.0"
description = ""
authors = ["Ryo Yamamoto <ryochaaaan0414@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^2.2.17"
psycopg2 = "^2.8.6"
pytest = "^6.1.2"
pytest-django = "^4.1.0"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
